agent-playground/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
colorama>=0.4.6
click>=8.0.0
